Move the tasks for GitHub releases to separate package, only put one class in one Kotlin file
