+++ src/test/kotlin/github/ReleasesSpecTest.kt
@@ -1,12 +1,8 @@
-
+package org.openstreetmap.josm.gradle.plugin.github
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertNull
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
-import org.openstreetmap.josm.gradle.plugin.task.DEFAULT_PICKUP_RELEASE_DESCRIPTION
-import org.openstreetmap.josm.gradle.plugin.task.DEFAULT_PICKUP_RELEASE_LABEL
-import org.openstreetmap.josm.gradle.plugin.task.DEFAULT_PICKUP_RELEASE_NAME
-import org.openstreetmap.josm.gradle.plugin.task.ReleasesSpec
 
 class ReleasesSpecTest {
